checksum linking

+++ src/jtl/Connector/Linker/ChecksumLinker.php
@@ -38,13 +38,14 @@ class ChecksumLinker
                     && $model->getId()->getEndpoint() !== null && strlen($model->getId()->getEndpoint() > 0)) {
                     $checksum->setEndpoint(self::$loader->read($model->getId()->getEndpoint(), $checksum->getType()));
 
-                    if ($checksum->getEndpoint() !== null) {
+                    if ($checksum->getEndpoint() !== null && strlen($checksum->getEndpoint()) > 0) {
                         if (($checksum->getEndpoint() !== $checksum->getHost())) {
                             $checksum->setHasChanged(true);
                             self::$loader->delete($model->getId()->getEndpoint(), $checksum->getType());
                             self::$loader->write($model->getId()->getEndpoint(), $checksum->getType(), $checksum->getHost());
                         }
                     } else {
+                        $checksum->setHasChanged(true);
                         self::$loader->write($model->getId()->getEndpoint(), $checksum->getType(), $checksum->getHost());
                     }
                 }
